+++ src/ydsh.cpp
@@ -580,9 +580,10 @@ unsigned int DSState_complete(DSState *st, DSCompletionOp op, unsigned int index
       ref = StringRef(*value, index);
     }
     auto old = st->getGlobal(BuiltinVarOffset::EXIT_STATUS);
-    unsigned int size = doCodeCompletion(*st, resolveModuleFromDesc(*st, ""), ref);
+    auto ret = doCodeCompletion(*st, "", ref);
+    assert(ret.hasValue());
     st->setGlobal(BuiltinVarOffset::EXIT_STATUS, std::move(old));
-    return size;
+    return ret.unwrap();
   }
   case DS_COMP_GET:
     if (value == nullptr) {